aws_sdk_ssooidc
Module client

Client for calling AWS SSO OIDC.

Using the Client

A client has a function for every operation that can be performed by the service. For example, the CreateToken operation has a Client::create_token, function which returns a builder for that operation. The fluent builder ultimately has a send() function that returns an async future that returns a result, as illustrated below:

let result = client.create_token()
    .client_id("example")
    .send()
    .await;

The underlying HTTP requests that get made by this can be modified with the customize_operation function on the fluent builder. See the customize module for more information.

Modules

customize

Operation customization and supporting types.

Structs

Client

Client for AWS SSO OIDC