Module aws_sdk_ssooidc::error

Modules

access_denied_exception

See AccessDeniedException

authorization_pending_exception

See AuthorizationPendingException

expired_token_exception

See ExpiredTokenException

internal_server_exception

See InternalServerException

invalid_client_exception

See InvalidClientException

invalid_client_metadata_exception

See InvalidClientMetadataException

invalid_grant_exception

See InvalidGrantException

invalid_request_exception

See InvalidRequestException

invalid_scope_exception

See InvalidScopeException

slow_down_exception

See SlowDownException

unauthorized_client_exception

See UnauthorizedClientException

unsupported_grant_type_exception

See UnsupportedGrantTypeException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AuthorizationPendingException

Indicates that a request to authorize a client with an access user session token is pending.

CreateTokenError

ExpiredTokenException

Indicates that the token issued by the service is expired and is no longer valid.

InternalServerException

Indicates that an error from the service occurred while trying to process a request.

InvalidClientException

Indicates that the clientId or clientSecret in the request is

InvalidClientMetadataException

Indicates that the client information sent in the request during registration is invalid.

InvalidGrantException

Indicates that a request contains an invalid grant. This can occur if a client makes a CreateToken request with an invalid grant type.

InvalidRequestException

Indicates that something is wrong with the input to the request. For example, a required

InvalidScopeException

Indicates that the scope provided in the request is invalid.

RegisterClientError

SlowDownException

Indicates that the client is making the request too frequently and is more than the service can handle.

StartDeviceAuthorizationError

UnauthorizedClientException

Indicates that the client is not currently authorized to make the request. This can happen

UnsupportedGrantTypeException

Indicates that the grant type in the request is not supported by the service.

Enums

CreateTokenErrorKind

RegisterClientErrorKind

StartDeviceAuthorizationErrorKind