Enum aws_sdk_ssooidc::operation::create_token_with_iam::CreateTokenWithIAMError

#[non_exhaustive]
pub enum CreateTokenWithIAMError {

    AccessDeniedException(AccessDeniedException),
    AuthorizationPendingException(AuthorizationPendingException),
    ExpiredTokenException(ExpiredTokenException),
    InternalServerException(InternalServerException),
    InvalidClientException(InvalidClientException),
    InvalidGrantException(InvalidGrantException),
    InvalidRequestException(InvalidRequestException),
    InvalidRequestRegionException(InvalidRequestRegionException),
    InvalidScopeException(InvalidScopeException),
    SlowDownException(SlowDownException),
    UnauthorizedClientException(UnauthorizedClientException),
    UnsupportedGrantTypeException(UnsupportedGrantTypeException),
    Unhandled(Unhandled),

}

Error type for the CreateTokenWithIAMError operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

You do not have sufficient access to perform this action.

AuthorizationPendingException(AuthorizationPendingException)

Indicates that a request to authorize a client with an access user session token is pending.

ExpiredTokenException(ExpiredTokenException)

Indicates that the token issued by the service is expired and is no longer valid.

InternalServerException(InternalServerException)

Indicates that an error from the service occurred while trying to process a request.

InvalidClientException(InvalidClientException)

Indicates that the clientId or clientSecret in the request is invalid. For example, this can occur when a client sends an incorrect clientId or an expired clientSecret.

InvalidGrantException(InvalidGrantException)

Indicates that a request contains an invalid grant. This can occur if a client makes a CreateToken request with an invalid grant type.

InvalidRequestException(InvalidRequestException)

Indicates that something is wrong with the input to the request. For example, a required parameter might be missing or out of range.

InvalidRequestRegionException(InvalidRequestRegionException)

Indicates that a token provided as input to the request was issued by and is only usable by calling IAM Identity Center endpoints in another region.

InvalidScopeException(InvalidScopeException)

Indicates that the scope provided in the request is invalid.

SlowDownException(SlowDownException)

Indicates that the client is making the request too frequently and is more than the service can handle.

UnauthorizedClientException(UnauthorizedClientException)

Indicates that the client is not currently authorized to make the request. This can happen when a clientId is not issued for a public client.

UnsupportedGrantTypeException(UnsupportedGrantTypeException)

Indicates that the grant type in the request is not supported by the service.

Unhandled(Unhandled)

๐Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): ยยยerr if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).

Implementations

impl CreateTokenWithIAMError

pub fn unhandled(err: impl Into<Box<dyn Error + Send + Sync + 'static>>) -> Self

Creates the CreateTokenWithIAMError::Unhandled variant from any error type.

pub fn generic(err: ErrorMetadata) -> Self

Creates the CreateTokenWithIAMError::Unhandled variant from an ErrorMetadata.

pub fn meta(&self) -> &ErrorMetadata

Returns error metadata, which includes the error code, message, request ID, and potentially additional information.

pub fn is_access_denied_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::AccessDeniedException.

pub fn is_authorization_pending_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::AuthorizationPendingException.

pub fn is_expired_token_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::ExpiredTokenException.

pub fn is_internal_server_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InternalServerException.

pub fn is_invalid_client_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InvalidClientException.

pub fn is_invalid_grant_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InvalidGrantException.

pub fn is_invalid_request_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InvalidRequestException.

pub fn is_invalid_request_region_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InvalidRequestRegionException.

pub fn is_invalid_scope_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::InvalidScopeException.

pub fn is_slow_down_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::SlowDownException.

pub fn is_unauthorized_client_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::UnauthorizedClientException.

pub fn is_unsupported_grant_type_exception(&self) -> bool

Returns true if the error kind is CreateTokenWithIAMError::UnsupportedGrantTypeException.