Module aws_sdk_ssooidc::types::error

Error types that AWS SSO OIDC can respond with.

Modules

builders

Builders

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AuthorizationPendingException

Indicates that a request to authorize a client with an access user session token is pending.

ExpiredTokenException

Indicates that the token issued by the service is expired and is no longer valid.

InternalServerException

Indicates that an error from the service occurred while trying to process a request.

InvalidClientException

Indicates that the clientId or clientSecret in the request is invalid. For example, this can occur when a client sends an incorrect clientId or an expired clientSecret.

InvalidClientMetadataException

Indicates that the client information sent in the request during registration is invalid.

InvalidGrantException

Indicates that a request contains an invalid grant. This can occur if a client makes a CreateToken request with an invalid grant type.

InvalidRequestException

Indicates that something is wrong with the input to the request. For example, a required parameter might be missing or out of range.

InvalidRequestRegionException

Indicates that a token provided as input to the request was issued by and is only usable by calling IAM Identity Center endpoints in another region.

InvalidScopeException

Indicates that the scope provided in the request is invalid.

SlowDownException

Indicates that the client is making the request too frequently and is more than the service can handle.

UnauthorizedClientException

Indicates that the client is not currently authorized to make the request. This can happen when a clientId is not issued for a public client.

UnsupportedGrantTypeException

Indicates that the grant type in the request is not supported by the service.