Struct aws_sdk_ssooidc::types::error::ExpiredTokenException

#[non_exhaustive]
pub struct ExpiredTokenException {
    pub error: Option<String>,
    pub error_description: Option<String>,
    pub message: Option<String>,
    /* private fields */
}

Indicates that the token issued by the service is expired and is no longer valid.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

error: Option<String>

Single error code. For this exception the value will be expired_token.

error_description: Option<String>

Human-readable text providing additional information, used to assist the client developer in understanding the error that occurred.

message: Option<String>
Implementations

impl ExpiredTokenException

pub fn error(&self) -> Option<&str>

Single error code. For this exception the value will be expired_token.

pub fn error_description(&self) -> Option<&str>

Human-readable text providing additional information, used to assist the client developer in understanding the error that occurred.

impl ExpiredTokenException

pub fn message(&self) -> Option<&str>

Returns the error message.

impl ExpiredTokenException

pub fn builder() -> ExpiredTokenExceptionBuilder

Creates a new builder-style object to manufacture ExpiredTokenException.