Module aws_sdk_ssooidc::config

Configuration for AWS SSO OIDC.

Modules

endpoint

Types needed to configure endpoint resolution.

http

HTTP request and response types.

interceptors

Types needed to implement Intercept.

retry

Retry configuration.

timeout

Timeout configuration.

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

BehaviorVersion

Behavior version of the client

Builder

Builder for creating a Config.

Config

Configuration for a aws_sdk_ssooidc service client.

ConfigBag

Layered configuration structure

Credentials

AWS SDK Credentials

FrozenLayer

FrozenLayer is the immutable and shareable form of Layer.

IdentityCache

Identity cache configuration.

Layer

A named layer comprising a config bag

Region

The region to send requests to.

RuntimeComponents

Components that can only be set in runtime plugins that the orchestrator uses directly to call an operation.

RuntimeComponentsBuilder

Builder for RuntimeComponents.

SharedAsyncSleep

Wrapper type for sharable AsyncSleep

SharedCredentialsProvider

Credentials Provider wrapper that may be shared

SharedHttpClient

Shared HTTP client for use across multiple clients and requests.

SharedIdentityCache

Shared identity cache.

SharedInterceptor

Interceptor wrapper that may be shared

SharedRuntimePlugin

Shared runtime plugin

Sleep

Future returned by AsyncSleep.

StalledStreamProtectionConfig

Configuration for stalled stream protection.

Traits

AsyncSleep

Async trait with a sleep function.

HttpClient

HTTP client abstraction.

Intercept

An interceptor allows injecting code into the SDK ’s request execution pipeline.

IntoShared

Like the Into trait, but for (efficiently) converting into a shared type.

ProvideCredentials

Asynchronous Credentials Provider

ResolveCachedIdentity

Caching resolver for identities.

RuntimePlugin

Runtime plugin trait